requirements.txt
kafka-python>=2.0.2
streamlit>=1.41.0
pandas>=2.3.0
plotly>=5.24.0
numpy>=2.1.0